+++ src/variables/ShellCommandVariable_FileName.ts
@@ -6,7 +6,7 @@ export class ShellCommandVariable_FileName extends ShellCommandVariable{
     getValue(): string {
         let file = this.app.workspace.getActiveFile();
         if (!file) {
-            this.newError("No file is active at the moment. Open a file or click a pane that has a file open.");
+            this.newErrorMessage("No file is active at the moment. Open a file or click a pane that has a file open.");
             return null; // null indicates that getting a value has failed and the command should not be executed.
         }
         return file.name;