+++ src/variables/ShellCommandVariable_FilePath.ts
@@ -15,11 +15,11 @@ export class ShellCommandVariable_FilePath extends ShellCommandVariable{
                 case "relative":
                     return active_file.path;
                 default:
-                    this.newError(`Unknown mode "${mode}"! Use "absolute" or "relative".`);
+                    this.newErrorMessage(`Unknown mode "${mode}"! Use "absolute" or "relative".`);
                     return null; // null indicates that getting a value has failed and the command should not be executed.
             }
         } else {
-            this.newError("No file is active at the moment. Open a file or click a pane that has a file open.");
+            this.newErrorMessage("No file is active at the moment. Open a file or click a pane that has a file open.");
             return null; // null indicates that getting a value has failed and the command should not be executed.
         }
     }