+++ src/variables/ShellCommandVariable_Title.ts
@@ -8,7 +8,7 @@ export class ShellCommandVariable_Title extends ShellCommandVariable{
         if (active_file) {
             return active_file.basename;
         }
-        this.newError("No file is active at the moment. Open a file or click a pane that has a file open.")
+        this.newErrorMessage("No file is active at the moment. Open a file or click a pane that has a file open.")
         return null;
     }
 }