+++ src/variables/parseShellCommandVariables.ts
@@ -11,22 +11,28 @@ import {ShellCommandVariable_FolderPath} from "./ShellCommandVariable_FolderPath
 import {ShellCommandVariable_Title} from "./ShellCommandVariable_Title";
 import {ShellCommandVariable_Workspace} from "./ShellCommandVariable_Workspace";
 
-export function parseShellCommandVariables(plugin: ShellCommandsPlugin, command: string, enable_error_messages: boolean) {
+/**
+ * @param plugin
+ * @param command
+ * @return string|string[] If parsing fails, an array of string error messages is returned. If the parsing succeeds, the parsed shell command will be returned just as a string, not in an array.
+ */
+export function parseShellCommandVariables(plugin: ShellCommandsPlugin, command: string): string | string[] {
     let shell_variables: ShellCommandVariable[] = [
-        new ShellCommandVariable_Clipboard(plugin, enable_error_messages),
-        new ShellCommandVariable_Date(plugin, enable_error_messages),
-        new ShellCommandVariable_FileName(plugin, enable_error_messages),
-        new ShellCommandVariable_FilePath(plugin, enable_error_messages),
-        new ShellCommandVariable_FolderName(plugin, enable_error_messages),
-        new ShellCommandVariable_FolderPath(plugin, enable_error_messages),
-        new ShellCommandVariable_Selection(plugin, enable_error_messages),
-        new ShellCommandVariable_Title(plugin, enable_error_messages),
-        new ShellCommandVariable_VaultPath(plugin, enable_error_messages),
-        new ShellCommandVariable_Workspace(plugin, enable_error_messages),
+        new ShellCommandVariable_Clipboard(plugin),
+        new ShellCommandVariable_Date(plugin),
+        new ShellCommandVariable_FileName(plugin),
+        new ShellCommandVariable_FilePath(plugin),
+        new ShellCommandVariable_FolderName(plugin),
+        new ShellCommandVariable_FolderPath(plugin),
+        new ShellCommandVariable_Selection(plugin),
+        new ShellCommandVariable_Title(plugin),
+        new ShellCommandVariable_VaultPath(plugin),
+        new ShellCommandVariable_Workspace(plugin),
     ];
     let parsed_command = command; // Create a copy of the variable because we don't want to alter the original value of 'command' during iterating its regex matches.
-    let parsing_failed = false;
-    shell_variables.forEach((variable: ShellCommandVariable) => {
+    for (let variable_index in shell_variables)
+    {
+        let variable: ShellCommandVariable = shell_variables[variable_index];
         let pattern = new RegExp(variable.getPattern(), "ig"); // i: case-insensitive; g: match all occurrences instead of just the first one.
         let match;
         while ((match = pattern.exec(command)) !== null) {
@@ -37,19 +43,16 @@ export function parseShellCommandVariables(plugin: ShellCommandsPlugin, command:
                 argument = match[1];
             }
             let variable_value = variable.getValue(argument);
-            if (null === variable_value) {
-                // The variable value getter has indicated that there has been a problem (probably in the passed argument) and executing the command should be cancelled. No need to continue iterating other variables.
-                parsing_failed = true;
-                // TODO: Find out how to break from forEach() so that we don't need to unnecessarily iterate the rest of the variables.
+            if (variable.getErrorMessages().length) {
+                // There has been a problem and executing the command should be cancelled.
+                console.log("Parsing command " + command + " failed.");
+                return variable.getErrorMessages(); // Returning now prevents parsing rest of the variables.
             }
             else
             {
                 parsed_command = parsed_command.replace(substitute, variable_value);
             }
         }
-    });
-    if (parsing_failed) {
-        return null;
     }
     return parsed_command;
 }